% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Create tibble containing distances between states i and j for all combinations of (i, j)}
\usage{
create_distance_matrix()
}
\value{
= no return value
}
\description{
Create tibble containing distances between states i and j for all combinations of (i, j)
}
