% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_historical.R
\name{create_historical_matrix}
\alias{create_historical_matrix}
\title{Create matrix from the historical data}
\usage{
create_historical_matrix(hist_data)
}
\arguments{
\item{hist_data}{= tibble containing historical data}
}
\value{
= 51x51 matrix of historical data
}
\description{
Create matrix from the historical data
}
