% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{generate_gravity_matrices}
\alias{generate_gravity_matrices}
\title{Generate a gravity model migration matrix for each column in the population estimates tibble
(used in combination with map() for more efficient file generation)}
\usage{
generate_gravity_matrices(col_name, data, alpha, locations)
}
\arguments{
\item{col_name}{= string, name of a column in the population estimates tibble (ex: "year_2009")}

\item{data}{= the population estimates tibble (pop_estimates)}

\item{alpha}{= hyperparameter used in calculations (can be fine tuned using historical data)}

\item{locations}{= tibble containing distances between each (i, j) state combination}
}
\value{
= a list of migration matrices calculated for each column of data in pop_estimates
}
\description{
Generate a gravity model migration matrix for each column in the population estimates tibble
(used in combination with map() for more efficient file generation)
}
