% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{generate_int_opp_files}
\alias{generate_int_opp_files}
\title{Generate intervening opportunity files}
\usage{
generate_int_opp_files(col_name, data, locations)
}
\arguments{
\item{col_name}{= string, name of a column in the population estimates tibble (ex: "year_2009")}

\item{data}{= the population estimates tibble (pop_estimates)}

\item{locations}{= tibble containing distances between each (i, j) state combination}
}
\value{
= no return value
}
\description{
Generate intervening opportunity files
}
