% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{generate_radiation_matrices}
\alias{generate_radiation_matrices}
\title{Generate a radiation model migration matrix for each column in the population estimates tibble
(used in combination with map() for more efficient file generation)}
\usage{
generate_radiation_matrices(col_name, data, alpha)
}
\arguments{
\item{col_name}{= string, name of a column in the population estimates tibble (ex: "year_2009")}

\item{data}{= the population estimates tibble (pop_estimates)}

\item{alpha}{= hyperparameter used in calculations (can be fine tuned using historical data)}
}
\value{
= a list of migration matrices calculated for each column of data in pop_estimates
}
\description{
Generate a radiation model migration matrix for each column in the population estimates tibble
(used in combination with map() for more efficient file generation)
}
