% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_gravity_prediction}
\alias{get_gravity_prediction}
\title{Run the gravity model and return the estimated population for the following year}
\usage{
get_gravity_prediction(pop_data, year, gravity_alpha, locations)
}
\arguments{
\item{pop_data}{= a vector containing population estimates for each state for a specific year (states listed in fips order)}

\item{year}{= the specific year of data which will be used in the name of the resulting csv file}

\item{gravity_alpha}{= historically tuned hyperparameter value}

\item{locations}{= tibble containing distances between each (i, j) state combination}
}
\value{
= a vector containing predicted population estimates for the following year given the gravity model's prediction
}
\description{
Run the gravity model and return the estimated population for the following year
}
