% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{get_intervening_opps}
\alias{get_intervening_opps}
\title{Calculate an intervening opportunities matrix and write it out as a csv}
\usage{
get_intervening_opps(locations, year_data, fips, filename)
}
\arguments{
\item{locations}{= tibble containing distances between each (i, j) state combination}

\item{year_data}{= a vector containing population estimates for each state for a specific year (states listed in fips order)}

\item{fips}{= list of states' FIPS id}

\item{filename}{= filename to use when writing data out to a csv}
}
\value{
= no return value
}
\description{
Calculate an intervening opportunities matrix and write it out as a csv
}
