% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{get_radiation_prediction}
\alias{get_radiation_prediction}
\title{Run the radiation model and return the estimated population for the following year}
\usage{
get_radiation_prediction(
  pop_data,
  s_ij_file,
  new_s_ij_file,
  year,
  radiation_alpha
)
}
\arguments{
\item{pop_data}{= a vector containing population estimates for each state for a specific year (states listed in fips order)}

\item{s_ij_file}{= string, path to where appropriate intervening opportunities csv can be found}

\item{new_s_ij_file}{= string, name to use for a newly calculated intervening opportunities csv file}

\item{year}{= the specific year of data which will be used in the name of the resulting radiation matrix csv file}

\item{radiation_alpha}{= historically tuned hyperparameter value}
}
\value{
= a vector containing predicted population estimates for the following year given the radiation model's prediction
}
\description{
Run the radiation model and return the estimated population for the following year
}
