% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{graph_population_change}
\alias{graph_population_change}
\title{Graph predicted population change 2018-2022}
\usage{
graph_population_change(pop_over_time, state_name, model)
}
\arguments{
\item{pop_over_time}{= tibble containing predicted population counts for all states for 2018-2022}

\item{state_name}{= the name of the state you want to see the population change graph for}

\item{model}{= name of the model you used (Gravity of Radiation)}
}
\value{
= ggplot line graph of predicted population change in the state between 2018-2022
}
\description{
Graph predicted population change 2018-2022
}
