% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_to_historical.R
\name{gravity_heatmap}
\alias{gravity_heatmap}
\title{Create gravity model residuals heatmap}
\usage{
gravity_heatmap(grav_matrix, hist_matrix, year)
}
\arguments{
\item{grav_matrix}{= 51x51 gravity model migration matrix}

\item{hist_matrix}{= 51x51 historical data matrix}

\item{year}{= string, year to label the heatmap}
}
\value{
= ggplot visual of a heatmap
}
\description{
Create gravity model residuals heatmap
}
