% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_models.R
\name{radiation_model}
\alias{radiation_model}
\title{Radiation model to produce a migration matrix (a prediction)}
\usage{
radiation_model(pop_data, intervening_data, names, alpha)
}
\arguments{
\item{pop_data}{= a vector containing population estimates for each state for a specific year (states listed in fips order)}

\item{intervening_data}{= string, path to where appropriate intervening opportunities csv can be found}

\item{names}{= a list of states' FIPS id}

\item{alpha}{= hyperparameter used in calculations (can be fine tuned using historical data)}
}
\value{
= migration matrix with values estimating number of migrants from states i to j for all combinations of (i, j)
}
\description{
Radiation model to produce a migration matrix (a prediction)
}
