% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{run_gravity_model}
\alias{run_gravity_model}
\title{Run the gravity model and return the sum of squared residuals
(used in combination with optimize() to find the most appropriate alpha value)}
\usage{
run_gravity_model(alpha, year_data, hist_data_file, locations)
}
\arguments{
\item{alpha}{= hyperparameter used in calculations}

\item{year_data}{= a vector containing population estimates for each state for a specific year (states listed in fips order)}

\item{hist_data_file}{= string, path to the appropriate csv containing historical data on the recorded number of migrants}

\item{locations}{= tibble containing distances between each (i, j) state combination}
}
\value{
= the sum of squared residuals (residual = recorded num_migrants - gravity model prediction)
}
\description{
Run the gravity model and return the sum of squared residuals
(used in combination with optimize() to find the most appropriate alpha value)
}
